// Logitude/AngularModules/CypressTesting/Maintenance/cypress/models/SignatureDetails.ts
export interface SignatureDetails {
  Date: string
  User: string
  Signature: string
  Logo: string
  SmallLogo: string
  WideLogo: string
  LocalCurrency: string
  Company: string
  Email: string
  Website: string
  IATA: string
  VATNo: string
  AddressID: string
  Contact: string
  SupporteMail: string
  UserSignatureImage: string
}